+++ integration_tests/test_upgrade.py
@@ -16,7 +16,6 @@ from .utils import (
     deploy_contract,
     edit_ini_sections,
     get_consensus_params,
-    get_send_enable,
     send_transaction,
     wait_for_block,
     wait_for_new_blocks,
@@ -27,13 +26,8 @@ pytestmark = pytest.mark.upgrade
 
 
 @pytest.fixture(scope="module")
-def testnet(tmp_path_factory):
+def custom_cronos(tmp_path_factory):
     yield from setup_cronos_test(tmp_path_factory)
-
-
-@pytest.fixture(scope="module")
-def mainnet(tmp_path_factory):
-    yield from setup_cronos_test(tmp_path_factory, False)
 
 
 def init_cosmovisor(home):
@@ -67,11 +61,11 @@ def post_init(path, base_port, config):
     )
 
 
-def setup_cronos_test(tmp_path_factory, testnet=True):
+def setup_cronos_test(tmp_path_factory):
     path = tmp_path_factory.mktemp("upgrade")
-    port = 26100 if testnet else 26200
-    nix_name = "upgrade-testnet-test-package" if testnet else "upgrade-test-package"
-    cfg_name = "cosmovisor_testnet" if testnet else "cosmovisor"
+    port = 26200
+    nix_name = "upgrade-test-package"
+    cfg_name = "cosmovisor"
     cmd = [
         "nix-build",
         Path(__file__).parent / f"configs/{nix_name}.nix",
@@ -91,7 +85,7 @@ def setup_cronos_test(tmp_path_factory, testnet=True):
         yield cronos
 
 
-def exec(c, tmp_path_factory, testnet=True):
+def exec(c, tmp_path_factory):
     """
     - propose an upgrade and pass it
     - wait for it to happen
@@ -103,16 +97,14 @@ def exec(c, tmp_path_factory, testnet=True):
         {"denom": "basetcro", "enabled": False},
         {"denom": "stake", "enabled": True},
     ]
-    p = cli.query_bank_send() if testnet else get_send_enable(port)
+    p = cli.query_bank_send()
     assert sorted(p, key=lambda x: x["denom"]) == send_enable
 
     # export genesis from old version
     c.supervisorctl("stop", "all")
     migrate = tmp_path_factory.mktemp("migrate")
     file_path0 = Path(migrate / "old.json")
-    with open(file_path0, "w") as fp:
-        json.dump(json.loads(cli.export()), fp)
-        fp.flush()
+    cli.export(output_document=str(file_path0))
 
     c.supervisorctl("start", "cronos_777-1-node0", "cronos_777-1-node1")
     wait_for_port(ports.evmrpc_port(c.base_port(0)))
@@ -123,13 +115,13 @@ def exec(c, tmp_path_factory, testnet=True):
     print("upgrade height", target_height)
 
     w3 = c.w3
-
-    if not testnet:
-        # before upgrade, PUSH0 opcode is not supported
-        with pytest.raises(ValueError) as e_info:
-            deploy_contract(w3, CONTRACTS["Greeter"])
-        assert "invalid opcode: PUSH0" in str(e_info.value)
-
+    random_contract = deploy_contract(
+        c.w3,
+        CONTRACTS["Random"],
+    )
+    with pytest.raises(ValueError) as e_info:
+        random_contract.caller.randomTokenId()
+    assert "invalid memory address or nil pointer dereference" in str(e_info.value)
     contract = deploy_contract(w3, CONTRACTS["TestERC20A"])
     old_height = w3.eth.block_number
     old_balance = w3.eth.get_balance(ADDRS["validator"], block_identifier=old_height)
@@ -139,7 +131,7 @@ def exec(c, tmp_path_factory, testnet=True):
     )
     print("old values", old_height, old_balance, old_base_fee)
 
-    plan_name = "v1.1.0-testnet-1" if testnet else "v1.1.0"
+    plan_name = "v1.1"
     rsp = cli.gov_propose_legacy(
         "community",
         "software-upgrade",
@@ -150,10 +142,10 @@ def exec(c, tmp_path_factory, testnet=True):
             "upgrade-height": target_height,
             "deposit": "10000basetcro",
         },
-        mode=None if testnet else "block",
+        mode=None,
     )
     assert rsp["code"] == 0, rsp["raw_log"]
-    approve_proposal(c, rsp, event_query_tx=testnet)
+    approve_proposal(c, rsp, event_query_tx=True)
 
     # update cli chain binary
     c.chain_binary = (
@@ -181,9 +173,11 @@ def exec(c, tmp_path_factory, testnet=True):
     )
     assert receipt.status == 1
 
-    if not testnet:
-        # after upgrade, PUSH0 opcode is supported
-        deploy_contract(w3, CONTRACTS["Greeter"])
+    # deploy contract should still work
+    deploy_contract(w3, CONTRACTS["Greeter"])
+    # random should work
+    res = random_contract.caller.randomTokenId()
+    assert res > 0, res
 
     # query json-rpc on older blocks should success
     assert old_balance == w3.eth.get_balance(
@@ -208,21 +202,6 @@ def exec(c, tmp_path_factory, testnet=True):
     assert rsp["params"]["min_timeout_duration"] == "3600s", rsp
     max_callback_gas = cli.query_params()["max_callback_gas"]
     assert max_callback_gas == "50000", max_callback_gas
-    if not testnet:
-        # migrate to sdk v0.47
-        c.supervisorctl("stop", "all")
-        sdk_version = "v0.47"
-        file_path1 = Path(migrate / f"{sdk_version}.json")
-        with open(file_path1, "w") as fp:
-            json.dump(cli.migrate_sdk_genesis(sdk_version, str(file_path0)), fp)
-            fp.flush()
-        # migrate to cronos v1.0.x
-        cronos_version = "v1.0"
-        file_path0 = Path(migrate / f"{cronos_version}.json")
-        with open(file_path0, "w") as fp:
-            json.dump(cli.migrate_cronos_genesis(cronos_version, str(file_path1)), fp)
-            fp.flush()
-        print(cli.validate_genesis(str(file_path0)))
 
     # update the genesis time = current time + 5 secs
     newtime = datetime.utcnow() + timedelta(seconds=5)
@@ -239,9 +218,5 @@ def exec(c, tmp_path_factory, testnet=True):
     c.supervisorctl("stop", "all")
 
 
-def test_cosmovisor_upgrade_mainnet(mainnet: Cronos, tmp_path_factory):
-    exec(mainnet, tmp_path_factory, False)
-
-
-def test_cosmovisor_upgrade_testnet(testnet: Cronos, tmp_path_factory):
-    exec(testnet, tmp_path_factory, True)
+def test_cosmovisor_upgrade(custom_cronos: Cronos, tmp_path_factory):
+    exec(custom_cronos, tmp_path_factory)